Fixed untangling logic
1. Fixed issue, when current revision had both NODE_PATH and
NODE_COPYFROM_PATH but NODE_COPYFROM_REV pointed to the previously
dropped revision that had it's NODE_PATH in excluded. Now all dropped
revisions are stored in the "skipped revisions" list and current
revision's NODE_COPYFROM_REV is checked against it. If NODE_COPYFROM_REV
is in the list, then revision is untangled.

2. Fixed issue, when revision was untangled when there were no need for
it. It lead to the wrong version of file being stored in the node record
and subsequent "svnadmin: E200014: Copy source checksum mismatch" error
on dump import. This occured when node record had  NODE_COPYFROM_REV and
NODE_COPYFROM_PATH, but also had TEXT_CONTENT_LEN. Now those revisions
are not untangled and NODE_COPYFROM_REV and NODE_COPYFROM_PATH are
stripped from them. This case is mentioned in the Subversion dumpfile
format description
(https://svn.apache.org/repos/asf/subversion/trunk/notes/dump-load-format.txt):

It is even possible to have a replace with a copyfrom source *and*
text, such as would result from this on the client side:

-------------------------------------------------------------------
$ svn rm dir/file.txt
$ svn cp otherdir/otherfile.txt dir/file.txt
$ echo "Replacement text" > dir/file.txt
$ svn ci -m "Replace dir/file.txt with a copy of otherdir/otherfile.txt
and replace its text, too."
-------------------------------------------------------------------